Refactor subscription screens to handle 2 flows
