+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/usersubscription/ui/UiElements.kt
@@ -1,13 +1,10 @@
 package io.horizontalsystems.bankwallet.modules.usersubscription.ui
 
-import androidx.activity.compose.LocalActivity
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -33,7 +30,6 @@ import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.SheetState
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -45,19 +41,12 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.modules.usersubscription.BuySubscriptionChoosePlanViewModel
-import io.horizontalsystems.bankwallet.modules.usersubscription.BuySubscriptionModel.badge
 import io.horizontalsystems.bankwallet.modules.usersubscription.BuySubscriptionModel.descriptionStringRes
 import io.horizontalsystems.bankwallet.modules.usersubscription.BuySubscriptionModel.iconRes
-import io.horizontalsystems.bankwallet.modules.usersubscription.BuySubscriptionModel.stringRepresentation
-import io.horizontalsystems.bankwallet.modules.usersubscription.BuySubscriptionModel.title
 import io.horizontalsystems.bankwallet.modules.usersubscription.BuySubscriptionModel.titleStringRes
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.ButtonPrimaryDefaults
@@ -72,13 +61,9 @@ import io.horizontalsystems.bankwallet.ui.compose.components.body_bran
 import io.horizontalsystems.bankwallet.ui.compose.components.caption_grey
 import io.horizontalsystems.bankwallet.ui.compose.components.headline1_leah
 import io.horizontalsystems.bankwallet.ui.compose.components.subhead1_leah
-import io.horizontalsystems.bankwallet.ui.compose.components.subhead2_jacob
-import io.horizontalsystems.bankwallet.ui.compose.components.subhead2_remus
 import io.horizontalsystems.bankwallet.ui.extensions.BottomSheetHeader
-import io.horizontalsystems.subscriptions.core.HSPurchase
 import io.horizontalsystems.subscriptions.core.IPaidAction
 import io.horizontalsystems.subscriptions.core.Subscription
-import io.horizontalsystems.subscriptions.core.numberOfDays
 
 val yellowGradient = Brush.horizontalGradient(
     colors = listOf(
@@ -92,177 +77,7 @@ val steelBrush = Brush.horizontalGradient(
 )
 
 @Composable
-fun SelectSubscriptionBottomSheet(
-    subscriptionId: String,
-    onDismiss: () -> Unit,
-    viewModel: BuySubscriptionChoosePlanViewModel = viewModel(),
-    onPurchase: () -> Unit,
-    onError: (Throwable) -> Unit,
-) {
-    val uiState = viewModel.uiState
-    val activity = LocalActivity.current
-
-    LaunchedEffect(Unit) {
-        viewModel.getBasePlans(subscriptionId)
-    }
-
-    uiState.error?.let {
-        onError(it)
-        viewModel.onErrorHandled()
-    }
-
-    uiState.purchase?.let {
-        if (it.status == HSPurchase.Status.Purchased) {
-            onPurchase()
-        }
-    }
-
-    val selectedItemIndex = uiState.selectedIndex
-    val freeTrialPeriodDays = uiState.freeTrialPeriod?.let {
-        stringResource(R.string.Period_Days, it.numberOfDays())
-    }
-
-    BottomSheetHeader(
-        iconPainter = painterResource(R.drawable.ic_circle_clock_24),
-        iconTint = ColorFilter.tint(ComposeAppTheme.colors.jacob),
-        title = stringResource(R.string.Premium_SelectSubscription),
-        onCloseClick = onDismiss
-    ) {
-        Column(
-            modifier = Modifier.padding(horizontal = 16.dp, vertical = 12.dp),
-            verticalArrangement = Arrangement.SpaceBetween,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Column(
-                verticalArrangement = Arrangement.spacedBy(10.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                uiState.basePlans.forEachIndexed { index, basePlan ->
-                    SubscriptionOption(
-                        title = basePlan.title(),
-                        price = basePlan.stringRepresentation(),
-                        note = "",
-                        isSelected = selectedItemIndex == index,
-                        badgeText = basePlan.badge(),
-                        onClick = {
-                            viewModel.select(index)
-                        }
-                    )
-                }
-            }
-
-            val bottomText = if (freeTrialPeriodDays != null) {
-                buildAnnotatedString {
-                    withStyle(SpanStyle(color = ComposeAppTheme.colors.remus)) {
-                        append(
-                            text = stringResource(
-                                R.string.Premium_EnjoyFreePeriod,
-                                freeTrialPeriodDays
-                            )
-                        )
-                    }
-                    append(" ")
-                    withStyle(SpanStyle(color = ComposeAppTheme.colors.leah)) {
-                        append(text = stringResource(R.string.Premium_CancelSubscriptionInfo))
-                    }
-                }
-            } else {
-                buildAnnotatedString {
-                    withStyle(SpanStyle(color = ComposeAppTheme.colors.leah)) {
-                        append(text = stringResource(R.string.Premium_CancelSubscriptionInfo))
-                    }
-                }
-            }
-
-
-            VSpacer(12.dp)
-            Text(
-                text = bottomText,
-                color = ComposeAppTheme.colors.grey,
-                style = ComposeAppTheme.typography.subheadR,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.padding(horizontal = 32.dp, vertical = 12.dp)
-            )
-            VSpacer(24.dp)
-
-            val buttonTitle = if (freeTrialPeriodDays != null) {
-                stringResource(R.string.Premium_GetFreePeriod, freeTrialPeriodDays)
-            } else {
-                stringResource(R.string.Premium_Subscribe)
-            }
-            ButtonPrimaryYellow(
-                modifier = Modifier.fillMaxWidth(),
-                title = buttonTitle,
-                onClick = {
-                    activity?.let {
-                        viewModel.launchPurchaseFlow(
-                            subscriptionId = subscriptionId,
-                            offerToken = uiState.basePlans[selectedItemIndex].offerToken,
-                            activity = it
-                        )
-                    }
-                }
-            )
-            VSpacer(36.dp)
-        }
-    }
-}
-
-@Composable
-fun SubscriptionOption(
-    title: String,
-    price: String,
-    note: String,
-    isSelected: Boolean,
-    badgeText: String?,
-    onClick: () -> Unit
-) {
-    val borderColor =
-        if (isSelected) ComposeAppTheme.colors.jacob else ComposeAppTheme.colors.blade
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .border(0.5.dp, borderColor, shape = RoundedCornerShape(16.dp))
-            .clip(RoundedCornerShape(16.dp))
-            .clickable { onClick() }
-            .padding(horizontal = 16.dp, vertical = 12.dp)
-    ) {
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                headline1_leah(title)
-                if (badgeText != null) {
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Box(
-                        modifier = Modifier
-                            .background(
-                                ComposeAppTheme.colors.remus,
-                                shape = RoundedCornerShape(8.dp)
-                            )
-                            .padding(horizontal = 6.dp, vertical = 2.dp)
-                    ) {
-                        Text(
-                            text = badgeText,
-                            color = ComposeAppTheme.colors.blade,
-                            style = ComposeAppTheme.typography.microSB,
-                        )
-                    }
-                }
-            }
-
-            Row() {
-                subhead2_jacob(price)
-                if (note.isNotEmpty()) {
-                    HSpacer(4.dp)
-                    subhead2_remus(note)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun ColumnScope.PlanItems(
+fun PlanItems(
     items: List<IPaidAction>,
     onItemClick: (IPaidAction) -> Unit
 ) {
@@ -552,31 +367,5 @@ fun InfoBottomSheet(
                 VSpacer(32.dp)
             }
         }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun SubscriptionBottomSheet(
-    modalBottomSheetState: SheetState,
-    subscription: Subscription,
-    navController: NavController,
-    hideBottomSheet: () -> Unit,
-    onError: (Throwable) -> Unit
-) {
-    ModalBottomSheet(
-        onDismissRequest = hideBottomSheet,
-        sheetState = modalBottomSheetState,
-        containerColor = ComposeAppTheme.colors.transparent
-    ) {
-        SelectSubscriptionBottomSheet(
-            subscriptionId = subscription.id,
-            onDismiss = hideBottomSheet,
-            onPurchase = {
-                hideBottomSheet()
-                navController.navigate("premium_subscribed_page")
-            },
-            onError = onError
-        )
     }
 }
